Refactorize query params to filters
