+++ package.json
@@ -27,7 +27,8 @@
     "forest-express": "3.2.4",
     "http-errors": "1.7.2",
     "lodash": "4.17.13",
-    "moment": "2.24.0"
+    "moment": "2.24.0",
+    "semver": "5.6.0"
   },
   "devDependencies": {
     "babel-cli": "7.0.0-beta.3",
@@ -46,7 +47,6 @@
     "mongoose-fixture-loader": "1.0.2",
     "onchange": "6.0.0",
     "pre-commit": "1.2.2",
-    "semver": "5.6.0",
     "simple-git": "1.65.0"
   },
   "scripts": {
